feat: add AIToggleButton component and integrate it into MailLayout; refactor AI chat handling in MailList
