+++ apps/mail/components/mail/mail-list.tsx
@@ -50,7 +50,6 @@ import { useParams, useRouter } from 'next/navigation';
 import { useTRPC } from '@/providers/query-provider';
 import { useThreadLabels } from '@/hooks/use-labels';
 import { useKeyState } from '@/hooks/use-hot-key';
-import { useIsMobile } from '@/hooks/use-mobile';
 import { useSession } from '@/lib/auth-client';
 import { RenderLabels } from './render-labels';
 import { Badge } from '@/components/ui/badge';
@@ -1007,55 +1006,9 @@ export const MailList = memo(({ isCompact }: MailListProps) => {
           <div className="h-4" />
         )}
       </div>
-
-      {/* Toggle AI Assistant Button - Only visible on md or smaller screens */}
-      <AIToggleButton />
     </>
   );
 });
-
-// AI Toggle Button Component
-const AIToggleButton = () => {
-  const { toggleOpen: toggleAISidebar, open: isSidebarOpen } = useAISidebar();
-
-  return (
-    !isSidebarOpen && (
-      <div className="fixed bottom-4 right-4 z-50">
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Button
-              variant="outline"
-              size="icon"
-              className="dark:bg-sidebar h-12 w-12 rounded-full"
-              onClick={(e) => {
-                e.stopPropagation();
-                toggleAISidebar();
-              }}
-            >
-              <div className="flex items-center justify-center">
-                <Image
-                  src="/black-icon.svg"
-                  alt="AI Assistant"
-                  width={22}
-                  height={22}
-                  className="block dark:hidden"
-                />
-                <Image
-                  src="/white-icon.svg"
-                  alt="AI Assistant"
-                  width={22}
-                  height={22}
-                  className="hidden dark:block"
-                />
-              </div>
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent>Toggle AI Assistant</TooltipContent>
-        </Tooltip>
-      </div>
-    )
-  );
-};
 
 MailList.displayName = 'MailList';
